+++ src/index.ts
@@ -4,3 +4,6 @@
 
 export * from './helpers';
 export * from './strict-builtins';
+export * from './loose-builtins';
+
+export * from './EXPERIMENTAL';