+++ src/loose-builtins/index.ts
@@ -1,0 +1,5 @@
+/**
+ * @file An entry point for loose versions of TS's built-in utility types.
+ */
+
+export * from './loose-omit';