+++ src/strict-builtins/index.ts
@@ -1,5 +1,5 @@
 /**
- * @file An entry point for strict version of TS's built-in utility types.
+ * @file An entry point for strict versions of TS's built-in utility types.
  */
 
 export * from './strict-omit';